Mockito/src/test/java/dio/bootcamp/santander/mockito/ContaTest.java
package dio.bootcamp.santander.mockito;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentMatchers;
import org.mockito.InOrder;
import org.mockito.Mockito;
import org.mockito.Spy;
import org.mockito.junit.jupiter.MockitoExtension;

/**
 *
 * @author José Gustavo
 */
@ExtendWith (MockitoExtension.class)
public class ContaTest {
    
    @Spy
    private Conta conta = new Conta(1_000);
    
    @Test
    void validarOrdemDeChamada(){
        conta.pagaBoleto(300);
        //ArgumentMatchers.anyInt();
        InOrder inOrder = Mockito.inOrder(conta);
        inOrder.verify(conta).pagaBoleto(300);
        inOrder.verify(conta).debita(300);
        inOrder.verify(conta).enviaCreditoParaEmissor(300);
    }
    
    @Test
    void validaQuantidadeDeVezesQueMétodoFoiChamado() {

        conta.validaSaldo(100);
        conta.validaSaldo(100);
        conta.validaSaldo(100);

        Mockito.verify(conta, Mockito.times(3)).validaSaldo(100);
    }
    
    @Test
    void naoFazNadaNaValidacaoDeSaldo(){
        Mockito.doNothing().when(conta).validaSaldo(ArgumentMatchers.anyInt());
        conta.validaSaldo(-1);
    }
    
}
